+++ projects/egeo/src/lib/st-pagination/st-pagination.interface.ts
@@ -31,3 +31,8 @@ export interface PaginateTextsSchema {
    perPage: TranslateableElement;
    of: TranslateableElement;
 }
+
+export interface PaginateIconClasses {
+   previous: string;
+   next: string;
+}